Fix auctions not cycling bidders correctly when quitting and skipping players
Last player used to be skipped because of a modulo off by one error. The bidders were also advanced when a bidder quit, causing the bidder after to be skipped.

Some changes were also made to better encapsulate the auction's responsibilities when it comes to tracking the balance of the current bidder and the current bidder itself.