Refactor buyable tiles behind interface
The class of actions for tiles that can be auctioned or bought has been moved to a Buyable interface. Tiles can still be homogenously aggregated and later bought because all tiles now implement an asBuyable method that returns an optional.

The optional is inhabited if the underlying tile type is buyable, otherwise, it is empty.